% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_final.R
\name{generate_all_data}
\alias{generate_all_data}
\title{Generate simulated data}
\usage{
generate_all_data(
  n = 200,
  alpha_S = 0,
  beta_M = 0,
  M.family = gaussian(),
  Y.family = gaussian(),
  sigma_M = 0.5,
  sigma_Y = 0.5
)
}
\arguments{
\item{n}{the sample size}

\item{alpha_S}{the parameter in M~S}

\item{beta_M}{the parameter in Y~S+M}

\item{M.family}{a description of the error distribution and link function to be used in the mediator model. This is the
result of a call to a family function. (See \code{\link{family}} for detials of family functions.) The default
family is gaussian().}

\item{Y.family}{a description of the error distribution and link function to be used in the outcome model. This is the
result of a call to a family function. (See \code{\link{family}} for detials of family functions.) The default
family is gaussian().}

\item{sigma_M}{the noise level for M~S}

\item{sigma_Y}{the noise level for Y~S+M}
}
\value{
A list with the following components:
\item{S}{exposure}
\item{M}{mediator}
\item{Y}{outcome}
\item{X}{confounder}
}
\description{
Generate simulated data
}
